feat(dev): handle skip ci label to skip tests

+++ .circleci/base-install.sh
@@ -7,10 +7,13 @@ cd "$DIR/.."
 # npm install just enough to run these scripts
 npm i --ignore-scripts --no-optional --only=prod --no-package-lock
 node .circleci/modules-to-test.js | tee packages/test.list
-./.circleci/assert-branch.sh
-./_scripts/create-version-json.sh
+if ([[ "$MODULE" == "many" ]] && grep -e '.' packages/test.list > /dev/null) ||
+    grep -e "$MODULE" -e 'all' packages/test.list > /dev/null; then
+  ./.circleci/assert-branch.sh
+  ./_scripts/create-version-json.sh
 
-sudo apt-get install -y graphicsmagick
+  sudo apt-get install -y graphicsmagick
 
-yarn install --immutable
-yarn workspace fxa-shared run build
+  yarn install --immutable
+  yarn workspace fxa-shared run build
+fi
